# Fix condition to align methods to 32 bytes boundary