In https://github.com/dotnet/runtime/pull/2249, we started doing alignment of methods to 32-byte boundary for `Tier1`. However, a method having loops bypass tiering and hence this condition was never executed. Fixed it to make sure we do the alignment for optimized methods only and don't do it for prejitted methods.