@kunalspathak 2020-10-06
> Any idea what a typical delta this causes on the overall size of the code segment?

I added logging around the change to see how many methods get 32-byte aligned vs. the default alignment and then ran pmi on .NET libraries. There is definitely duplicate methods that might get counted in either of them, but on an average, 532,667 methods continued getting default alignment while 70,573 methods got 32-bytes alignment i.e. out of 603,240 total methods, 70,573 methods got 32-byte aligned which is approx. **11%** of them. This is close to what @AndyAyersMS  anticipated in https://github.com/dotnet/runtime/pull/2249#issuecomment-579117528.

> LoopReturn seems like a reasonable choice, though I am also interested cases where the benchmark payload does not have a loop. Maybe we can find one?

I tried gathering the results for [Fib benchmark](https://github.com/dotnet/performance/blob/454476401e17ed7f4d8b899ecf7661eb6cd63bad/src/benchmarks/micro/runtime/Benchstones/BenchI/Fib.cs#L12) (the only one which is bimodel and has no-loops) and seeing mixed results.

master:

| Method | Mean     | Error   | StdDev  | Median   | Min      | Max      |
|--------|----------|---------|---------|----------|----------|----------|
|   Fib  | 136.0 us | 0.61 us | 0.51 us | 135.8 us | 135.3 us | 137.0 us |
| Fib    | 142.2 us | 0.55 us | 0.51 us | 142.2 us | 141.4 us | 143.4 us |
| Fib    | 142.3 us | 0.79 us | 0.70 us | 142.3 us | 141.5 us | 143.9 us |
| Fib    | 142.2 us | 0.70 us | 0.65 us | 142.0 us | 141.5 us | 143.6 us |
| Fib    | 143.2 us | 1.94 us | 1.72 us | 142.8 us | 141.5 us | 147.1 us |


PR:

| Method | Mean     | Error   | StdDev  | Median   | Min      | Max      |
|--------|----------|---------|---------|----------|----------|----------|
|   Fib  | 142.1 us | 0.82 us | 0.73 us | 141.9 us | 141.3 us | 144.1 us |
| Fib    | 142.0 us | 0.51 us | 0.45 us | 142.0 us | 141.4 us | 143.0 us |
| Fib    | 142.1 us | 0.54 us | 0.48 us | 142.3 us | 141.3 us | 142.9 us |
| Fib    | 142.2 us | 0.70 us | 0.62 us | 142.0 us | 141.5 us | 143.6 us |
| Fib    | 149.5 us | 3.73 us | 4.15 us | 147.7 us | 143.2 us | 158.1 us |