@kunalspathak 2020-10-06
For x86 as well, I don't see much difference for Fib benchmark on my local machine.

master:

| Method | Mean     | Error   | StdDev  | Median   | Min      | Max      |
|--------|----------|---------|---------|----------|----------|----------|
|   Fib  | 142.4 us | 1.04 us | 0.87 us | 142.4 us | 141.3 us | 144.2 us |
| Fib    | 145.7 us | 2.79 us | 2.99 us | 144.8 us | 142.5 us | 151.9 us |
| Fib    | 142.6 us | 0.91 us | 0.81 us | 142.5 us | 141.7 us | 144.5 us |
| Fib    | 142.7 us | 0.79 us | 0.70 us | 142.5 us | 141.4 us | 143.8 us |
| Fib    | 142.5 us | 0.53 us | 0.49 us | 142.4 us | 141.7 us | 143.4 us |

PR:

| Method | Mean     | Error   | StdDev  | Median   | Min      | Max      |
|--------|----------|---------|---------|----------|----------|----------|
|   Fib  | 142.3 us | 0.77 us | 0.72 us | 142.4 us | 141.3 us | 143.5 us |
| Fib    | 142.5 us | 0.96 us | 0.85 us | 142.0 us | 141.6 us | 144.2 us |
| Fib    | 143.7 us | 1.34 us | 1.25 us | 143.4 us | 141.7 us | 145.9 us |
| Fib    | 142.1 us | 0.45 us | 0.38 us | 142.1 us | 141.5 us | 142.9 us |
| Fib    | 144.3 us | 2.08 us | 1.94 us | 143.7 us | 142.1 us | 148.1 us |